refactor: move InMemoryTokenRepository to /examples, move Decorator types (and others) to /types
