+++ src/decorators/ChatCommand.decorator.ts
@@ -15,7 +15,7 @@ Przykład użycia:
 
 */
 
-import { GeneralContainer, GeneralFactory, GeneralRegistry, GeneralRegistryEntry } from "../storage/runtime/Decorator.storage";
+import { GeneralContainer, GeneralFactory, GeneralRegistry, GeneralRegistryEntry } from "../types/Decorator.storage.types";
 import { ChatCommandDecoratorOptions, ChatCommandExecution, ChatCommandInstance } from "../types/ChatCommand.types";
 import ChannelChatMessageEventData from "../types/EventSub_Events/ChannelChatMessageEventData.types";
 import {Logger} from "../utils/Logger";
